+++ R/params.R
@@ -1,3 +1,25 @@
+# Reverse g3_parameterized(), find out how parameter would be broken down
+g3_parameterized_breakdown <- function (in_c) {
+    if (!is.call(in_c)) stop("Expect a g3_parameterized() call")
+    fn_name <- deparse1(in_c[[1]])
+
+    # Ignore value modifiers
+    if (fn_name == '~') return(g3_parameterized_breakdown(in_c[[2]]))
+    if (fn_name == 'exp') return(g3_parameterized_breakdown(in_c[[2]]))
+    if (fn_name == '*') return(g3_parameterized_breakdown(in_c[[2]]))
+    if (fn_name == 'avoid_zero') return(g3_parameterized_breakdown(in_c[[2]]))
+    if (fn_name == '+') return(g3_parameterized_breakdown(in_c[[2]]))
+
+    if (fn_name == 'stock_prepend') return(c("stock", g3_parameterized_breakdown(in_c[[3]])))
+    if (fn_name == 'g3_param') return(as.character(c()))
+    if (fn_name == 'g3_param_table') {
+        if (deparse1(in_c[[3]][[1]]) != "expand.grid") stop("Expect g3_param_table() to use expand.grid()")
+        col_names <- names(in_c[[3]])
+        return(col_names[nzchar(col_names)])
+    }
+    stop("Unknown parameter definition: ", deparse1(in_c))
+}
+
 g3_parameterized <- function(
         name,
         by_stock = FALSE,
@@ -105,7 +127,7 @@ g3_parameterized <- function(
 
     # Modify value if asked
     if (exponentiate) out <- substitute(exp(x), list(x = out))
-    if (scale != 1) out <- substitute(scale * x, list(x = out, scale = scale))
+    if (scale != 1) out <- substitute(x * scale, list(x = out, scale = scale))
     if (avoid_zero != 0) out <- substitute(avoid_zero(x), list(x = out))
     if (offset != 0) out <- substitute(x + offset, list(x = out, offset = offset))
     return(out)