+++ R/run_r.R
@@ -168,8 +168,7 @@ g3_to_r <- function(actions, trace = FALSE, strict = FALSE) {
     out <- call("function", pairlist(param = alist(y=)$y), as.call(c(
         list(as.symbol(open_curly_bracket)),
         scope,
-        all_actions_code,
-        quote(stop("Should have return()ed somewhere in the loop")))))
+        all_actions_code )))
 
     # Rework any g3_* function calls into the code we expect
     g3_functions <- function (in_code) {
@@ -231,4 +230,18 @@ edit.g3_r <- function(name = NULL, file = "", title = NULL, editor = getOption("
     attributes(out) <- oldattr
     environment(out) <- environment(name)
     return(out)
+}
+
+print.g3_r <- function(x, ..., with_environment = FALSE, with_template = FALSE) {
+    a <- attributes(x)
+    attributes(x) <- NULL
+    print.function(x)
+    if (with_environment) {
+        writeLines("Environment:")
+        str(as.list(environment(x)), no.list = TRUE)
+    }
+    if (with_template) {
+        writeLines("Parameter template:")
+        str(a$parameter_template, no.list = TRUE)
+    }
 }