+++ NEWS.md
@@ -12,6 +12,9 @@
 * Defaults for ``g3_suitability_exponentiall50``
 * Support cut() formatted character columns as well as factor columns in likelihood observation data.
 * Add ``g3_init_val()`` helper, to replace ``gadgetutils::g3_init_guess()``
+* Add ``g3_areas()`` helper to create numbered area vectors
+* Add ``g3_distribution_preview()`` helper to see how observatation data will be converted to array
+* ``period`` now optional for ``g3l_random_walk()``, ``g3l_random_dnorm()``
 
 ## Breaking changes